[TEST] Extract common test fixtures (#124)
