+++ test/graaflib/algorithm/coloring/greedy_coloring_test.cpp
@@ -1,12 +1,9 @@
 #include <graaflib/algorithm/coloring/greedy_graph_coloring.h>
-#include <graaflib/graph.h>
-#include <graaflib/types.h>
 #include <gtest/gtest.h>
+#include <utils/fixtures/fixtures.h>
 
 #include <unordered_map>
 #include <unordered_set>
-#include <utility>
-#include <vector>
 
 namespace graaf::algorithm {
 
@@ -18,10 +15,7 @@ struct TypedColoringTest : public testing::Test {
   using edge_t = typename T::second_type;
 };
 
-using graph_types =
-    testing::Types<directed_graph<int, int>, undirected_graph<int, int>>;
-
-TYPED_TEST_SUITE(GraphColoringTest, graph_types);
+TYPED_TEST_SUITE(GraphColoringTest, utils::fixtures::minimal_graph_types);
 
 }  // namespace
 
@@ -30,10 +24,7 @@ struct GreedyGraphColoringTest : public testing::Test {
   using graph_t = T;
 };
 
-using graph_types =
-    testing::Types<directed_graph<int, int>, undirected_graph<int, int>>;
-
-TYPED_TEST_SUITE(GreedyGraphColoringTest, graph_types);
+TYPED_TEST_SUITE(GreedyGraphColoringTest, utils::fixtures::minimal_graph_types);
 
 // Test case for an empty graph
 TYPED_TEST(GreedyGraphColoringTest, EmptyGraph) {
