+++ test/graaflib/algorithm/shortest_path/bellman_ford_test.cpp
@@ -1,20 +1,10 @@
 #include <graaflib/algorithm/shortest_path/bellman_ford.h>
 #include <gtest/gtest.h>
+#include <utils/fixtures/fixtures.h>
 
 namespace graaf::algorithm {
 
 namespace {
-
-template <typename T>
-class my_weighted_edge : public weighted_edge<T> {
- public:
-  explicit my_weighted_edge(T weight) : weight_{weight} {}
-
-  [[nodiscard]] T get_weight() const noexcept override { return weight_; }
-
- private:
-  T weight_{};
-};
 
 template <typename T>
 struct BellmanFordShortestPathsTest : public testing::Test {
@@ -22,76 +12,17 @@ struct BellmanFordShortestPathsTest : public testing::Test {
   using edge_t = typename T::second_type;
 };
 
-using weighted_graph_types = testing::Types<
-
-    /**
-     * Primitive edge type directed graph
-     */
-    std::pair<directed_graph<int, int>, int>,
-    std::pair<directed_graph<int, unsigned long>, unsigned long>,
-    std::pair<directed_graph<int, float>, float>,
-    std::pair<directed_graph<int, long double>, long double>,
-
-    /**
-     * Non primitive weighted edge type directed graph
-     */
-
-    std::pair<directed_graph<int, my_weighted_edge<int>>,
-              my_weighted_edge<int>>,
-    std::pair<directed_graph<int, my_weighted_edge<unsigned long>>,
-              my_weighted_edge<unsigned long>>,
-    std::pair<directed_graph<int, my_weighted_edge<float>>,
-              my_weighted_edge<float>>,
-    std::pair<directed_graph<int, my_weighted_edge<long double>>,
-              my_weighted_edge<long double>>,
-
-    /**
-     * Primitive edge type undirected graph
-     */
-    std::pair<undirected_graph<int, int>, int>,
-    std::pair<undirected_graph<int, unsigned long>, unsigned long>,
-    std::pair<undirected_graph<int, float>, float>,
-    std::pair<undirected_graph<int, long double>, long double>,
-
-    /**
-     * Non primitive weighted edge type undirected graph
-     */
-    std::pair<undirected_graph<int, my_weighted_edge<int>>,
-              my_weighted_edge<int>>,
-    std::pair<undirected_graph<int, my_weighted_edge<unsigned long>>,
-              my_weighted_edge<unsigned long>>,
-    std::pair<undirected_graph<int, my_weighted_edge<float>>,
-              my_weighted_edge<float>>,
-    std::pair<undirected_graph<int, my_weighted_edge<long double>>,
-              my_weighted_edge<long double>>>;
-
-TYPED_TEST_SUITE(BellmanFordShortestPathsTest, weighted_graph_types);
+TYPED_TEST_SUITE(BellmanFordShortestPathsTest,
+                 utils::fixtures::weighted_graph_types);
 
 template <typename T>
 struct BellmanFordShortestPathsSignedTypesTest : public testing::Test {
   using graph_t = typename T::first_type;
   using edge_t = typename T::second_type;
 };
-using directed_weighted_graph_signed_types = testing::Types<
-    /**
-     * Primitive edge type directed graph
-     */
-    std::pair<directed_graph<int, int>, int>,
-    std::pair<directed_graph<int, float>, float>,
-    std::pair<directed_graph<int, long double>, long double>,
-
-    /**
-     * Non primitive weighted edge type directed graph
-     */
-    std::pair<directed_graph<int, my_weighted_edge<int>>,
-              my_weighted_edge<int>>,
-    std::pair<directed_graph<int, my_weighted_edge<float>>,
-              my_weighted_edge<float>>,
-    std::pair<directed_graph<int, my_weighted_edge<long double>>,
-              my_weighted_edge<long double>>>;
 
 TYPED_TEST_SUITE(BellmanFordShortestPathsSignedTypesTest,
-                 directed_weighted_graph_signed_types);
+                 utils::fixtures::directed_weighted_graph_signed_types);
 
 }  // namespace
