+++ test/graaflib/algorithm/shortest_path/bfs_shortest_path_test.cpp
@@ -1,7 +1,6 @@
 #include <graaflib/algorithm/shortest_path/bfs_shortest_path.h>
-#include <graaflib/graph.h>
-#include <graaflib/types.h>
 #include <gtest/gtest.h>
+#include <utils/fixtures/fixtures.h>
 
 namespace graaf::algorithm {
 
@@ -11,9 +10,7 @@ struct TypedShortestPathTest : public testing::Test {
   using graph_t = T;
 };
 
-using graph_types =
-    testing::Types<directed_graph<int, int>, undirected_graph<int, int>>;
-TYPED_TEST_SUITE(TypedShortestPathTest, graph_types);
+TYPED_TEST_SUITE(TypedShortestPathTest, utils::fixtures::minimal_graph_types);
 }  // namespace
 
 TYPED_TEST(TypedShortestPathTest, BfsMinimalShortestPath) {
