+++ test/graaflib/algorithm/shortest_path/dijkstra_shortest_path_test.cpp
@@ -1,21 +1,11 @@
 #include <fmt/core.h>
 #include <graaflib/algorithm/shortest_path/dijkstra_shortest_path.h>
 #include <gtest/gtest.h>
+#include <utils/fixtures/fixtures.h>
 
 namespace graaf::algorithm {
 
 namespace {
-
-template <typename T>
-class my_weighted_edge : public weighted_edge<T> {
- public:
-  explicit my_weighted_edge(T weight) : weight_{weight} {}
-
-  [[nodiscard]] T get_weight() const noexcept override { return weight_; }
-
- private:
-  T weight_{};
-};
 
 template <typename T>
 struct DijkstraShortestPathTest : public testing::Test {
@@ -23,50 +13,8 @@ struct DijkstraShortestPathTest : public testing::Test {
   using edge_t = typename T::second_type;
 };
 
-using weighted_graph_types = testing::Types<
-
-    /**
-     * Primitive edge type directed graph
-     */
-    std::pair<directed_graph<int, int>, int>,
-    std::pair<directed_graph<int, unsigned long>, unsigned long>,
-    std::pair<directed_graph<int, float>, float>,
-    std::pair<directed_graph<int, long double>, long double>,
-
-    /**
-     * Non primitive weighted edge type directed graph
-     */
-
-    std::pair<directed_graph<int, my_weighted_edge<int>>,
-              my_weighted_edge<int>>,
-    std::pair<directed_graph<int, my_weighted_edge<unsigned long>>,
-              my_weighted_edge<unsigned long>>,
-    std::pair<directed_graph<int, my_weighted_edge<float>>,
-              my_weighted_edge<float>>,
-    std::pair<directed_graph<int, my_weighted_edge<long double>>,
-              my_weighted_edge<long double>>,
-
-    /**
-     * Primitive edge type undirected graph
-     */
-    std::pair<undirected_graph<int, int>, int>,
-    std::pair<undirected_graph<int, unsigned long>, unsigned long>,
-    std::pair<undirected_graph<int, float>, float>,
-    std::pair<undirected_graph<int, long double>, long double>,
-
-    /**
-     * Non primitive weighted edge type undirected graph
-     */
-    std::pair<undirected_graph<int, my_weighted_edge<int>>,
-              my_weighted_edge<int>>,
-    std::pair<undirected_graph<int, my_weighted_edge<unsigned long>>,
-              my_weighted_edge<unsigned long>>,
-    std::pair<undirected_graph<int, my_weighted_edge<float>>,
-              my_weighted_edge<float>>,
-    std::pair<undirected_graph<int, my_weighted_edge<long double>>,
-              my_weighted_edge<long double>>>;
-
-TYPED_TEST_SUITE(DijkstraShortestPathTest, weighted_graph_types);
+TYPED_TEST_SUITE(DijkstraShortestPathTest,
+                 utils::fixtures::weighted_graph_types);
 
 template <typename T>
 struct DijkstraShortestPathSignedTypesTest : public testing::Test {
@@ -74,45 +22,8 @@ struct DijkstraShortestPathSignedTypesTest : public testing::Test {
   using edge_t = typename T::second_type;
 };
 
-using weighted_graph_signed_types = testing::Types<
-
-    /**
-     * Primitive edge type directed graph
-     */
-    std::pair<directed_graph<int, int>, int>,
-    std::pair<directed_graph<int, float>, float>,
-    std::pair<directed_graph<int, long double>, long double>,
-
-    /**
-     * Non primitive weighted edge type directed graph
-     */
-
-    std::pair<directed_graph<int, my_weighted_edge<int>>,
-              my_weighted_edge<int>>,
-    std::pair<directed_graph<int, my_weighted_edge<float>>,
-              my_weighted_edge<float>>,
-    std::pair<directed_graph<int, my_weighted_edge<long double>>,
-              my_weighted_edge<long double>>,
-
-    /**
-     * Primitive edge type undirected graph
-     */
-    std::pair<undirected_graph<int, int>, int>,
-    std::pair<undirected_graph<int, float>, float>,
-    std::pair<undirected_graph<int, long double>, long double>,
-
-    /**
-     * Non primitive weighted edge type undirected graph
-     */
-    std::pair<undirected_graph<int, my_weighted_edge<int>>,
-              my_weighted_edge<int>>,
-    std::pair<undirected_graph<int, my_weighted_edge<float>>,
-              my_weighted_edge<float>>,
-    std::pair<undirected_graph<int, my_weighted_edge<long double>>,
-              my_weighted_edge<long double>>>;
-
 TYPED_TEST_SUITE(DijkstraShortestPathSignedTypesTest,
-                 weighted_graph_signed_types);
+                 utils::fixtures::weighted_graph_signed_types);
 
 }  // namespace
