+++ test/graaflib/algorithm/strongly_connected_components/tarjan_test.cpp
@@ -1,9 +1,7 @@
 #include <graaflib/algorithm/strongly_connected_components/tarjan.h>
-#include <graaflib/graph.h>
-#include <graaflib/types.h>
 #include <gtest/gtest.h>
+#include <utils/fixtures/fixtures.h>
 
-#include <algorithm>
 #include <vector>
 
 namespace graaf::algorithm {
@@ -14,23 +12,10 @@ struct TypedStronglyConnectedComponentsTest : public testing::Test {
   using graph_t = T;
 };
 
-using graph_types =
-    testing::Types<directed_graph<int, int>, undirected_graph<int, int>>;
-
-TYPED_TEST_SUITE(TypedStronglyConnectedComponentsTest, graph_types);
+TYPED_TEST_SUITE(TypedStronglyConnectedComponentsTest,
+                 utils::fixtures::minimal_directed_graph_type);
 
 }  // namespace
-
-template <typename T>
-class my_weighted_edge : public weighted_edge<T> {
- public:
-  explicit my_weighted_edge(T weight) : weight_{weight} {}
-
-  [[nodiscard]] T get_weight() const noexcept override { return weight_; }
-
- private:
-  T weight_{};
-};
 
 template <typename T>
 struct TarjansStronglyConnectedComponentsTest : public testing::Test {
@@ -60,30 +45,8 @@ bool are_set_vectors_equal(const std::vector<std::vector<vertex_id_t>>& vec1,
   return sorted_vec1 == sorted_vec2;
 }
 
-using weighted_graph_types = testing::Types<
-
-    /**
-     * Primitive edge type directed graph
-     */
-    std::pair<directed_graph<int, int>, int>,
-    std::pair<directed_graph<int, unsigned long>, unsigned long>,
-    std::pair<directed_graph<int, float>, float>,
-    std::pair<directed_graph<int, long double>, long double>,
-
-    /**
-     * Non primitive weighted edge type directed graph
-     */
-
-    std::pair<directed_graph<int, my_weighted_edge<int>>,
-              my_weighted_edge<int>>,
-    std::pair<directed_graph<int, my_weighted_edge<unsigned long>>,
-              my_weighted_edge<unsigned long>>,
-    std::pair<directed_graph<int, my_weighted_edge<float>>,
-              my_weighted_edge<float>>,
-    std::pair<directed_graph<int, my_weighted_edge<long double>>,
-              my_weighted_edge<long double>>>;
-
-TYPED_TEST_SUITE(TarjansStronglyConnectedComponentsTest, weighted_graph_types);
+TYPED_TEST_SUITE(TarjansStronglyConnectedComponentsTest,
+                 utils::fixtures::directed_weighted_graph_types);
 
 TYPED_TEST(TarjansStronglyConnectedComponentsTest,
            TarjansSingleVertexGraphTest) {
